Simplifying mtu chain elements
