+++ pkg/networkservice/mechanisms/vxlan/mtu/common.go
@@ -16,61 +16,6 @@
 
 package mtu
 
-import (
-	"context"
-	"io"
-	"net"
-
-	"git.fd.io/govpp.git/api"
-	"github.com/pkg/errors"
-
-	interfaces "github.com/edwarnicke/govpp/binapi/interface"
-	"github.com/edwarnicke/govpp/binapi/ip"
-
-	"github.com/networkservicemesh/sdk-vpp/pkg/tools/types"
-)
-
-func getMTU(ctx context.Context, vppConn api.Connection, tunnelIP net.IP) (uint32, error) {
-	client, err := interfaces.NewServiceClient(vppConn).SwInterfaceDump(ctx, &interfaces.SwInterfaceDump{})
-	if err != nil {
-		return 0, errors.Wrapf(err, "error attempting to get interface dump client to determine MTU for tunnelIP %q", tunnelIP)
-	}
-	defer func() { _ = client.Close() }()
-
-	for {
-		details, err := client.Recv()
-		if err == io.EOF {
-			break
-		}
-		if err != nil {
-			return 0, errors.Wrapf(err, "error attempting to get interface details to determine MTU for tunnelIP %q", tunnelIP)
-		}
-
-		ipAddressClient, err := ip.NewServiceClient(vppConn).IPAddressDump(ctx, &ip.IPAddressDump{
-			SwIfIndex: details.SwIfIndex,
-			IsIPv6:    tunnelIP.To4() == nil,
-		})
-		if err != nil {
-			return 0, errors.Wrapf(err, "error attempting to get ip address for vpp interface %q determine MTU for tunnelIP %q", details.InterfaceName, tunnelIP)
-		}
-		defer func() { _ = ipAddressClient.Close() }()
-
-		for {
-			ipAddressDetails, err := ipAddressClient.Recv()
-			if err == io.EOF {
-				break
-			}
-			if err != nil {
-				return 0, errors.Wrapf(err, "error attempting to get interface ip address for %q (swIfIndex: %q) to determine MTU for tunnelIP %q", details.InterfaceName, details.SwIfIndex, tunnelIP)
-			}
-			if types.FromVppAddressWithPrefix(ipAddressDetails.Prefix).IP.Equal(tunnelIP) && details.Mtu[0] != 0 {
-				return (details.Mtu[0] - overhead(tunnelIP.To4() == nil)), nil
-			}
-		}
-	}
-	return 0, errors.Errorf("unable to find interface in vpp with tunnelIP: %q or interface IP MTU is zero", tunnelIP)
-}
-
 func overhead(isV6 bool) uint32 {
 	if !isV6 {
 		// outer ipv4 header - 20 bytes
