+++ pkg/networkservice/mechanisms/wireguard/mtu/client.go
@@ -18,9 +18,8 @@ package mtu
 
 import (
 	"context"
+	"errors"
 	"net"
-	"sync"
-	"sync/atomic"
 
 	"git.fd.io/govpp.git/api"
 	"google.golang.org/grpc"
@@ -30,15 +29,14 @@ import (
 	"github.com/networkservicemesh/api/pkg/api/networkservice/mechanisms/wireguard"
 
 	"github.com/networkservicemesh/sdk/pkg/networkservice/core/next"
+	"github.com/networkservicemesh/sdk/pkg/networkservice/utils/metadata"
+
+	"github.com/networkservicemesh/sdk-vpp/pkg/networkservice/tunnelmtu"
 )
 
 type mtuClient struct {
 	vppConn  api.Connection
 	tunnelIP net.IP
-	mtu      uint32
-
-	inited    uint32
-	initMutex sync.Mutex
 }
 
 // NewClient - returns client chain element to manage wireguard MTU
@@ -50,15 +48,18 @@ func NewClient(vppConn api.Connection, tunnelIP net.IP) networkservice.NetworkSe
 }
 
 func (m *mtuClient) Request(ctx context.Context, request *networkservice.NetworkServiceRequest, opts ...grpc.CallOption) (*networkservice.Connection, error) {
-	if err := m.init(ctx); err != nil {
-		return nil, err
+	tunnelMTU, ok := tunnelmtu.Load(ctx, metadata.IsClient(m))
+	if !ok {
+		return nil, errors.New("tunnel MTU is required")
 	}
-	if mechanism := wireguard.ToMechanism(request.GetConnection().GetMechanism()); mechanism != nil && (mechanism.MTU() == 0 || mechanism.MTU() > m.mtu) {
-		mechanism.SetMTU(m.mtu)
+
+	mtu := tunnelMTU - overhead(m.tunnelIP.To4() == nil)
+	if mechanism := wireguard.ToMechanism(request.GetConnection().GetMechanism()); mechanism != nil && (mechanism.MTU() == 0 || mechanism.MTU() > mtu) {
+		mechanism.SetMTU(mtu)
 	}
 	for _, mech := range request.GetMechanismPreferences() {
-		if mechanism := wireguard.ToMechanism(mech); mechanism != nil && (mechanism.MTU() == 0 || mechanism.MTU() > m.mtu) {
-			mechanism.SetMTU(m.mtu)
+		if mechanism := wireguard.ToMechanism(mech); mechanism != nil && (mechanism.MTU() == 0 || mechanism.MTU() > mtu) {
+			mechanism.SetMTU(mtu)
 		}
 	}
 	conn, err := next.Client(ctx).Request(ctx, request, opts...)
@@ -70,22 +71,4 @@ func (m *mtuClient) Request(ctx context.Context, request *networkservice.Network
 
 func (m *mtuClient) Close(ctx context.Context, conn *networkservice.Connection, opts ...grpc.CallOption) (*emptypb.Empty, error) {
 	return next.Client(ctx).Close(ctx, conn, opts...)
-}
-
-func (m *mtuClient) init(ctx context.Context) error {
-	if atomic.LoadUint32(&m.inited) > 0 {
-		return nil
-	}
-	m.initMutex.Lock()
-	defer m.initMutex.Unlock()
-	if atomic.LoadUint32(&m.inited) > 0 {
-		return nil
-	}
-
-	var err error
-	m.mtu, err = getMTU(ctx, m.vppConn, m.tunnelIP)
-	if err == nil {
-		atomic.StoreUint32(&m.inited, 1)
-	}
-	return err
 }
