+++ pkg/networkservice/mechanisms/wireguard/mtu/doc.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2021 Cisco and/or its affiliates.
+// Copyright (c) 2021-2022 Cisco and/or its affiliates.
 //
 // SPDX-License-Identifier: Apache-2.0
 //
@@ -14,5 +14,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package mtu computes the mtu for the vxlan tunnel and adds it to the mechanism
+// Package mtu computes the mtu for the wireguard tunnel and adds it to the mechanism
 package mtu
