+++ pkg/networkservice/tunnelmtu/doc.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2020-2022 Cisco and/or its affiliates.
+// Copyright (c) 2022 Cisco and/or its affiliates.
 //
 // SPDX-License-Identifier: Apache-2.0
 //
@@ -14,26 +14,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package mtu
-
-func overhead(isV6 bool) uint32 {
-	// https://lists.zx2c4.com/pipermail/wireguard/2017-December/002201.html
-	if !isV6 {
-		//  20-byte outer IPv4 header
-		//  8-byte outer UDP header
-		//  4-byte type
-		//  4-byte key index
-		//  8-byte nonce
-		// 16-byte authentication tag
-		// 60 byte total
-		return 60
-	}
-	//  40-byte outer IPv4 header
-	//  8-byte outer UDP header
-	//  4-byte type
-	//  4-byte key index
-	//  8-byte nonce
-	// 16-byte authentication tag
-	// 80 bytes total
-	return 80
-}
+// Package tunnelmtu computes the mtu for the tunnels and adds it to the metadata
+package tunnelmtu
